add overwrite option to copy/move/rename files component

+++ WombatGH/CopyMoveRenameFile.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
-using System.Linq;
-using System.Text;
 using Grasshopper.Kernel;
 using WombatGH.Properties;
 
@@ -12,7 +9,8 @@ namespace WombatGH
     public class CopyMoveRenameFile : GH_Component
     {
 
-        public CopyMoveRenameFile() : base("Copy/Move/Rename Files", "CopyFile", "Allows you to copy or rename/move a file", "Wombat", "Files / Interop")
+        public CopyMoveRenameFile() : base("Copy/Move/Rename Files", "CopyFile", "Allows you to copy or rename/move a file",
+            "Wombat", "Files / Interop")
         {
         }
 
@@ -27,6 +25,8 @@ namespace WombatGH
             pManager.AddTextParameter("To Path", "T", "The path you want to move/copy/rename to", GH_ParamAccess.item);
             pManager.AddBooleanParameter("Move", "M",
                 "Set to true to move/rename the file, set to false to just make a copy", GH_ParamAccess.item, false);
+            pManager.AddBooleanParameter("Overwrite", "W",
+                "Set to true to overwrite the file in its original location", GH_ParamAccess.item, false);
         }
 
         protected override void RegisterOutputParams(GH_OutputParamManager pManager)
@@ -41,15 +41,17 @@ namespace WombatGH
             var fromPath = "";
             var toPath = "";
             var deleteOld = false;
+            var overwrite = false;
 
             if (!DA.GetData("Run", ref run)) return;
             if (!DA.GetData("From Path", ref fromPath)) return;
             if (!DA.GetData("To Path", ref toPath)) return;
             if (!DA.GetData("Move", ref deleteOld)) return;
+            if (!DA.GetData("Overwrite", ref overwrite)) return;
 
             if (!run) return;
 
-            File.Copy(fromPath,toPath);
+            File.Copy(fromPath,toPath, overwrite);
             if(deleteOld) File.Delete(fromPath);
 
             DA.SetData("Output path",toPath);
